+++ packages/language-support/src/antlr-grammar/CypherCmdLexer.g4
@@ -1,4 +1,4 @@
-lexer grammar CommandLexer;
+lexer grammar CypherCmdLexer;
 
 import CypherLexer;
