+++ packages/language-support/src/highlighting/syntaxColouring/syntaxColouringHelpers.ts
@@ -2,7 +2,7 @@ import { SemanticTokenTypes } from 'vscode-languageserver-types';
 
 import { Token } from 'antlr4';
 
-import CypherLexer from '../../generated-parser/CommandLexer';
+import CypherLexer from '../../generated-parser/CypherCmdLexer';
 
 import { CypherTokenType, lexerSymbols } from '../../lexerSymbols';
